+++ pdfsearcher/__init__.py
@@ -1,0 +1,2 @@
+import findPages 
+import searchText
